% Generated by roxygen2 (4.0.1): do not edit by hand
\name{D18}
\alias{D18}
\title{Compute D18 IBI}
\usage{
D18(data, all = FALSE)
}
\arguments{
\item{data}{User data merged with metadata, as generated by \code{\link{taxonInfo}}}

\item{all}{If false (default), return only SampleID, D18 IBI and QC. Else, all metrics are returned.}
}
\description{
Computes D18 IBI score and QC.
}
\examples{
data(algaeSamples)
preparedData <- taxonInfo(validator(algaeSamples))
D18(preparedData$cleanData)
}

