% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grabMetaData}
\alias{grabMetaData}
\title{Pull updated metadata}
\usage{
grabMetaData()
}
\value{
A data frame containing updated metadata. This will not replace the internal database in this package.
}
\description{
Pull updated metadata from Moss Landing Marine Labratory's data warehouse
}

