% Generated by roxygen2 (4.0.1): do not edit by hand
\name{H20}
\alias{H20}
\title{Compute H20 IBI}
\usage{
H20(data, all = FALSE)
}
\arguments{
\item{data}{User data merged with metadata, as generated by \code{\link{taxonInfo}}}

\item{all}{If false (default), return only SampleID, H20 IBI and QC. Else, all metrics are returned.}
}
\description{
Computes H20 IBI score and QC.
}
\examples{
data(algaeSamples)
preparedData <- taxonInfo(validator(algaeSamples))
H20(preparedData$cleanData)
}

