% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAlgaeMetaData.R
\name{loadAlgaeMetaData}
\alias{loadAlgaeMetaData}
\title{Read internal algae metadata table}
\usage{
loadAlgaeMetaData()
}
\description{
Read internal algae metadata table
}
\examples{
meta <- loadAlgaeMetaData()
View(meta)
}
