% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2.r
\name{S2}
\alias{S2}
\title{Compute S2 IBI}
\usage{
S2(data, all = FALSE)
}
\arguments{
\item{data}{User data merged with metadata, as generated by \code{\link{taxonInfo}}}

\item{all}{If false (default), return only SampleID, S2 IBI and QC. Else, all metrics are returned.}
}
\description{
Computes S2 IBI score and QC.
}
\examples{
data(algaeSamples)
preparedData <- taxonInfo(validator(algaeSamples))
S2(preparedData$cleanData)

}
