% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.r
\name{validator}
\alias{validator}
\title{Algae data validation}
\usage{
validator(data)
}
\description{
Checks that required columns are present in algae data frame
}
\examples{
data(algaeSamples)
validator(algaeSamples)
}
