% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonCheck.r
\name{taxonInfo}
\alias{taxonInfo}
\title{Algae taxon validation}
\usage{
taxonInfo(data, algaeMetaData = NULL)
}
\arguments{
\item{data}{A validated algae data frame}

\item{algaeMetaData}{A data frame contain taxonomic lookup data. If null (default), uses metadata distributed with this package}
}
\value{
A list containing the cleaned data, a report of the number rows purged per sample, a table listing all taxa in all samples,
and whether it was present in the metadata.
}
\description{
Checks that all FinalIDs are present in metadata. If not, the row is purged. Remainder are then merged together with metadata.
}
\examples{
data(algaeSamples)
tinfo <- taxonInfo(validator(algaeSamples))
tinfo$taxaSummary
}
